requirements.txt
autopep8==1.5.7
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
dataclasses==0.8
Flask==2.0.1
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.26
protobuf==3.18.1
pycodestyle==2.7.0
pytube==11.0.1
requests==2.26.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.6.0
